Add push stream manager.
The push stream manager is a per-origin cache of currently open push
streams. It serves as the URLProtocolClient in order to buffer lifetime
events like the response, data chunks, and finalization. New requests
coming from the app are checked against this cache, matching the
canonical URL. If a match is found, the stream is removed from the push
cache and appropriate URLProtocolClient catch-up callbacks are made.

The manager also maintains a mapping of associated streams to push
streams. If the original (associated) stream is cancelled, all
unclaimed, in-progress push streams are cancelled. The only way to
cancel a single in-progress push stream is to start a new NSURLRequest
to hook it up to the push stream, then cancel it.

Work inspired by @layerhq's push stream pull request. Big thanks to
@blakewatters and @chipxsd.